+++ tests/match_server/machine_c/layer1/game.py
@@ -11,7 +11,7 @@ class GameState(AbstractState):
     def name(self):
         return GAME
 
-    def exit(self, req):
+    def entry(self, req):
         """次の辺の名前を返します
 
         Parameters