+++ tests/match_server/machine_c/layer1/reply.py
@@ -11,7 +11,7 @@ class ReplyState(AbstractState):
     def name(self):
         return REPLY
 
-    def exit(self, req):
+    def entry(self, req):
         """次の辺の名前を返します
 
         Parameters